Separate views to view folder
