+++ app/views/events/show.vue
@@ -10,7 +10,7 @@
     <button type="button">Edit</button>
     <button type="button">Delete</button>
 
-    <list :token="csrfToken" :event="event" :orders="orders"></list>
+    <order-list :token="csrfToken" :event="event" :orders="orders"></order-list>
 
     <form method="post" :action="`/event/${eventId}/order/new`">
       <input type="hidden" name="_csrf" :value="csrfToken">
@@ -32,14 +32,14 @@
 </template>
 
 <script>
-  import list from './detail/components/list.vue';
+  import orderList from './events/_order-list.vue';
 
   export default {
     data: function () {
       return {};
     },
     components: {
-      list: list
+      orderList: orderList
     }
   };
 </script>
